Swapped KC_TAB and KC_BSPC.
Now, backspace is the left inner most thumb, and TAB is the right.
Why: vim snippet selecting TAB is often followed SPACE, which is next to
it and awkward.
Also, the comma, dot and SYMBL activation key are left hand.
Following a snippet selection with the right thumb press makes them easier to
press.