Change L_NUM_Y and L_NUM_W to KC_Y and KC_W
While getting used to the new physical key layout,
these pinky keys are failing with the current timeout setting.
That is, when typing 'y' or 'w' the associated layers are activated
but the chars are not printed. This happens often enough to introduce
the change. As the fingers get used, the change may be reverted.